pretrain-enc can eval

+++ dataloader/dataloader.py
@@ -13,7 +13,6 @@ class _BaseDataLoader(object):
 		# save pick_confirm for all pip indices
 		self.dataPipePath = './densecap/data_pipeline/'
 		self.pipeLen = 10
-		self.init_pick_confirm_files()
 
 	def updatePipeIndex(self):
 		self.pipeIndex = (self.pipeIndex+1)%self.pipeLen
@@ -26,7 +25,7 @@ class _BaseDataLoader(object):
 				with open(save_file,'w') as f:
 					f.write('a')
 
-	def loadOneJson(self):
+	def loadOneJson(self, mode):
 		'''
 		Every time loadJson load data for ONE image. The loaded data variable contains fields below:
 		# data['info'] = info[1]
@@ -45,7 +44,7 @@ class _BaseDataLoader(object):
 		while True:
 			filename = glob.glob(self.dataPipePath+'data_'+str(pInd)+'*')
 			if len(filename)>=1:
-				if not os.path.isfile(self.dataPipePath+'pick_confirm_'+str(pInd)):
+				if (not os.path.isfile(self.dataPipePath+'pick_confirm_'+str(pInd))) or mode=='test':
 					# pick_confirm file has been removed by lua loader. so this file is new. 
 					if not os.path.isfile(self.dataPipePath+'writing_block_'+str(pInd)):
 						# lua writing file finished.
@@ -63,10 +62,11 @@ class _BaseDataLoader(object):
 			reader = torchfile.T7Reader(f, utf8_decode_strings=True)
 			data = reader.read_obj()
 
-		os.remove(filename)
-		# place pick_confirm_$pipIndex to notify lua program
-		with open(self.dataPipePath+'pick_confirm_'+str(pInd), 'w') as f:
-			f.write('a')
+		if mode=='train':
+			os.remove(filename)
+			# place pick_confirm_$pipIndex to notify lua program
+			with open(self.dataPipePath+'pick_confirm_'+str(pInd), 'w') as f:
+				f.write('a')
 
 		# update pInd
 		self.updatePipeIndex()
@@ -77,10 +77,13 @@ class _BaseDataLoader(object):
 		raise NotImplementedError
 
 
-class LoaderEncTrain(_BaseDataLoader):
+class LoaderEnc(_BaseDataLoader):
 	"""docstring for LoaderEncTrain"""
-	def __init__(self):
-		super(LoaderEncTrain, self).__init__()
+	def __init__(self,mode='train'):
+		super(LoaderEnc, self).__init__()
+		if mode=='train':
+			self.init_pick_confirm_files()
+		self.mode = mode
 
 	def filtReplicate(self, data):
 		# shuffle 128 gt boxes copy
@@ -105,7 +108,7 @@ class LoaderEncTrain(_BaseDataLoader):
 		a batch of boxes and captions for ONE image. 
 		Every image has up to 128 boxes, numbers may vary. 
 		'''
-		data, itr, numiters = self.loadOneJson()
+		data, itr, numiters = self.loadOneJson(self.mode)
 		filtInds = self.filtReplicate(data)
 		# return only useful data fields
 		ret = {}
@@ -124,6 +127,7 @@ class _LoaderDec(_BaseDataLoader):
 	"""docstring for LoaderDec"""
 	def __init__(self, arg):
 		super(_LoaderDec, self).__init__()
+		self.init_pick_confirm_files()
 		self.arg = arg
 
 	def randomSample(self,data,boxes,sampleNum):
